Event listener registration can be chained.
The `on` method returns a reference to the emitter so calls can be chained. A
reference to the derived type is returned rather than an EventEmitter&. This
is to avoid accidental slicing e.g.
auto accidentalSlidedCopy = emitter.on<"name"_e>( ... )

To facilitate this, the EventEmitter template now needs to be passed the
derived type.